pyproject.toml
[project]
name = "piano-falling-notes"
version = "0.1.0"
description = "MusicXML to falling notes video generator (PianiCast style)"
requires-python = ">=3.10"
dependencies = [
    "music21>=9.0",
    "Pillow>=10.0",
    "numpy>=1.24",
    "tqdm>=4.60",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
]

[project.scripts]
piano-falling-notes = "piano_falling_notes.__main__:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
